requirements.txt
playwright
pytest-bdd
pytest-playwright
pytest-xdist